Extract eb instance, routingMap => environmentsByUrl
