+++ src/routingMap.ts
@@ -1,28 +1,13 @@
-import * as fs from 'fs';
-import * as path from 'path';
-import { ElasticBeanstalk } from 'aws-sdk';
-
+import { eb } from './eb';
 import { envNames } from './environments';
 
 import { setIsHealthy } from './health';
-
-const AWS_SECRET_PATH = path.join(process.cwd(), 'secrets/aws.json');
-
-const { accessKeyId, secretAccessKey } = JSON.parse(
-  fs.readFileSync(AWS_SECRET_PATH, 'utf8'),
-) as AwsSecrets;
-
-const eb = new ElasticBeanstalk({
-  region: 'us-east-1',
-  accessKeyId,
-  secretAccessKey,
-});
 
 process.on('unhandledRejection', () => {
   setIsHealthy(false);
 });
 
-export const routingMap = eb
+export const environmentsByUrl = eb
   .describeEnvironments({
     ApplicationName: 'hollowverse',
     EnvironmentNames: envNames,
